+++ tests/test_ds9_live.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python3
 import unittest
+import pyds9
 from pathlib import Path
 from bin import ds9_live
 
@@ -10,7 +11,7 @@ class TestDS9(unittest.TestCase):
     """
     def test_pyds9(self):
         print('pyds9.test is disabled because it is quite slow')
-        # pyds9.test()
+        pyds9.test()
 
     def test_no_args(self):
         args = ds9_live.parseargs()
@@ -20,13 +21,21 @@ class TestDS9(unittest.TestCase):
         window.close()
 
     def test_boss(self):
-        window = ds9_live.DS9Window('BOSS', '/data/spectro/', 'sdR-r1*',
-                                    'histequ', '1.0', False, False, True)
+        if Path('/data/spectro').exists():
+            window = ds9_live.DS9Window('BOSS', '/data/spectro/', 'sdR-r1*',
+                                        'histequ', '1.0', False, False, True)
+        else:
+            window = ds9_live.DS9Window('BOSS', '~/data/spectro/', 'sdR-r1*',
+                                        'histequ', '1.0', False, False, True)
         window.close()
 
     def test_guider(self):
-        window = ds9_live.DS9Window('Guider', '/data/gcam/', 'gimg-*',
-                                    'histequ', '1.0', False, False, True)
+        if Path('/data/gcam').exists():
+            window = ds9_live.DS9Window('Guider', '/data/gcam/', 'gimg-*',
+                                        'histequ', '1.0', False, False, True)
+        else:
+            window = ds9_live.DS9Window('Guider', '~/data/gcam/', 'gimg-*',
+                                        'histequ', '1.0', False, False, True)
         window.close()
 
     def test_apogee_summary_dir(self):